ai/flask/requirements.txt
# Gemini AI
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# Firebase
firebase-admin>=6.0.0

# Flask API
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=3.0.0
